BUG: recompute operator level info in parametrized instantiation
Otherwise the already computed weights of an operator can
rely on the arity of the operator before the dependence on
parameters of the definition (`Foo(x) == INSTANCE Inner`)
is added to the instantiated operator signature.

Parametric instantiation changes the arity of operators,
so the weights list needs to be recomputed, based on the
arity after instantiation has been performed.