TST: correctly detect end of module
This change ensures that the testing commands
are read after the closing horizontal rule (`=====*`)
of the module.

To do this, the nesting depth of submodules is tracked,
by incrementing the counter `submodule_nesting_depth` when
encountering a module opening (`-----*\s*MODULE`), and
decrementing this counter when encountering
a module closing (`=====*`).

Commands are read after the module's end.
Previously, commands were read after the first module closing
(`=====*`), which results in errors in the presence
of submodules.